Refactor so that workflows process input bindings (primarily secondaryFiles).
